update valid methods

+++ vpn_pool/db.py
@@ -45,18 +45,35 @@ class DB(object):
 
     def fetch_all(self):
         with self.connection() as con:
-            items = con.execute('select * from {}'.format(self.table_name))
-            return items.fetchall()
+            res = con.execute('select * from {}'.format(self.table_name))
+            return res.fetchall()
 
     # ids数据量过大时，一次删除会有 sqlite3.OperationalError: too many SQL variables 错误，应分批删除
     def delete(self,ids):
         with self.connection(True) as con:
             if isinstance(ids,list) or isinstance(ids,tuple) or isinstance(ids,set):
+                if len(ids)<1:
+                    return
                 if len(ids) > 100:
                     while len(ids) >0:
                         del_ids = ids[:100]
                         ids_param = (len(del_ids)*'?,')[:-1]
                         con.execute(('delete from {} where id in ('+ ids_param +')').format(self.table_name),tuple(del_ids))
                         ids = ids[100:]
+                else:
+                    ids_param = (len(ids) * '?,')[:-1]
+                    con.execute(('delete from {} where id in (' + ids_param + ')').format(self.table_name),tuple(ids))
             else:
                 con.execute('delete from {} where id =? '.format(self.table_name), (ids,))
+
+    def find_by_ip_port(self,ip,port):
+        with self.connection(True) as con:
+            res = con.execute('select * from {} where ip=? and port=? '.format(self.table_name),(ip,port))
+            return res.fetchall()
+
+    def update_stype(self,ids,stype):
+        if not ids:
+            return
+        params = [(stype,id) for id in ids]
+        with self.connection(True) as con:
+            con.executemany('update {} set stype=? where id=?'.format(self.table_name),params)
